Draft separating depth editor from call stack view
